chore(phpat): test shared infrastructure architecture
